fix(train): NaN stability - fp32 Newton-Schulz, finite loss guard
Changes:
- zeropower_via_newtonschulz5: use float32 instead of bfloat16 for stability
- train loop: skip backward step when loss is non-finite
- Prevents NaN cascades at high LR with fp16 autocast

+++ src/hagi/train/optim.py
@@ -15,7 +15,7 @@ def zeropower_via_newtonschulz5(G: torch.Tensor, steps: int = 5, eps: float = 1e
     """Approximate orthogonalization of a 2D matrix via quintic Newton-Schulz."""
     assert G.ndim == 2, "Muon orthogonalization expects a 2D matrix"
     a, b, c = (3.4445, -4.7750, 2.0315)
-    x = G.bfloat16()
+    x = G.float()
     transposed = G.size(0) > G.size(1)
     if transposed:
         x = x.T
@@ -26,7 +26,7 @@ def zeropower_via_newtonschulz5(G: torch.Tensor, steps: int = 5, eps: float = 1e
         x = a * x + B @ x
     if transposed:
         x = x.T
-    return x
+    return x.to(G.dtype)
 
 
 class Muon(torch.optim.Optimizer):
